WEBP-167: Checkboxes and Multiselect: remove copies of value prop in data
Vue seems to have problems with reactive arrays. Not really sure if it is realated to: https://vuejs.org/v2/guide/list.html#Caveats
but it seems to not update the $data.checkedValue when the value prop was updated, this only happens with arrays as props not singular types.

(WEBP-167: Frontend: rewrite mutliple selection widget (selection))