UIEXT-534: Refactor icon layouting in InputField
Instead of rendering icons on top of the input field, a flexbox layout
is now used to render icons before and after the input box (like it was
done in knime-ui for the SearchBar component).

Also, the InputField icon slots support the FunctionButton component and
hence no additional CSS is required in many cases to use buttons within
InputField.

UIEXT-534 (Add different filtering options to TwinList)