UIEXT-588: Introduce the concept of 're-usability' of a page
I.e. completely static pages can (and must) be marked as re-usable
if to be used accross different node- (and port-) implementations.

TableView- and DefaultNodeDialog-pages marked as re-usuable because the
former is used in the table view node, statistics node and port view;
while the latter is used in all nodes with a default node dialog.

Tests added and modified.

UIEXT-588 (Issues and refactoring related to the page-id)