UIEXT-534: improve handling of SVG icons with paths having fill color
So far the SVG style used for icons used with `FunctionButton` didn't
work for icons that use SVG paths that have a fill color set explicitly.
This however is done frequently for SVG paths that represent text. The
consequence was that these paths weren't colored correctly. This is
fixed now by explicitly setting the fill color for those SVG paths that
come with an non-empty fill attribute.

UIEXT-534 (Add different filtering options to TwinList)