AP-20007: Add configKeys to settings schema
by adding a `FieldNodeSettingsPersitor` interface that defines a
configKeys method. If a custom persistor is used, this method is used to
retrieve the keys that should be considered for showing or hiding the
flow variable indicator. Otherwise, the configKey of the `Persist`
annotation is used.

AP-20007 (Persistence layer can break flow variable overwrite indications)