NXT-1664: PortContext introduced
Available whenever a PortView or the data-services of a PortView are
created. It gives access to the actual port a port view is created for.
And the port in turn allows one to determine the actual native(!) node
the port data originates from (which in turn is required to create an
ExecutionContext as needed for the table view). Necessary for the case
a PortView is created for a metanode- or component-port.

NXT-1664 (Port view doesn't show up for component ports)