UIEXT-862: Introduce 'page-content-id' for non-static node-view pages
Because that page path needs to change on re-execution in case of
non-static node-views in order to make sure the browser updates the
respective page (and page resources) and doesn't take it from the
browser cache. And the new page-content-id in the path assures that.

Extra PagePathSegment-record introduced to represent (and document) the
page path segments more explicitly.

UIEXT-862 (Backend-rendered node views don't update on webportal page re-execution)