NXT-1811: PortViewEnt constructor changed
Helps to avoid problems of deleting the respective node (or closing the
workflow) while the PortViewEnt is being instantiated (which can take
some time, e.g., when calculating statistics which is delivered with the
initial data). With the change we avoid the creation of
PortView-instances again (because it's now passed in).

NXT-1811 (Use the very same thread for statistics calculation)