UIEXT-1345: Use renderer component instead of directly accessing the renderer
The renderer component is the renderer that gets used in the end. E.g.
when  having cells with svg and png images, the renderer itself is not
an AbstractPainterDataValueRenderer but the renderer that is used in the
end if it is a png image is one.

UIEXT-1345 (Images should not be scaled up)