UIEXT-1385: Use unique id for storing passwords and only store passwords in InitialData
This solves the issue that we need to send a new password of the same
field to the frontend when we overwrite the credentials field with a
flow variable. Since we do not want to apply this new password on
ApplyData, we should not store it in the first place.

UIEXT-1385 (Resolve Problems with CredentialsInput)