UIEXT-1535: Simplify initialization of Twinlist
The manually deselected were set incorrectly previously, which did not
matter since we do not need to set them at all, as they are set by the
onChange method afterwards anyway. This is also why the tests did not
catch that. Knowing that the outcome for manuallyDeselected do not matter at all
when processiung unknown columns, it is possible to simplify this
process quite a bit.

UIEXT-1535 (Error in console when opening TableView + Dialog)