UIEXT-1893: Intialize 'm_selected' in ColumnFilter
The effect is that the node dialog frontend containing
a column filter will also receive the initialized 'selected'-property.
And whenever dialog settings change and the updated settings are passed
to the view preview, those will also contain the 'selected'-property
(otherwise it wasn't present and caused the issue).

UIEXT-1893 (NPE in Sunburst Chart when changing any setting)