UIEXT-2001: Make update transformations inline
The initial idea of taking copies and returning them was meant to a)
simplify the code by limiting mutable state and b) allowing for settings
to be applied simultaneously after a cascase of updates.

but actually in retrospect having it inline makes the code more concise
(as it is now clear that the settings are to be manipulated and we do
not need to take deep copies anymore) and there is no problem with
having multiple data updates in a cascade of updates

Most of the changes in this commit is moving state to composabes and
refactoring the NodeDialog more to script setup, which will be continued
in the future.

UIEXT-2001 (Transitive ValueReference chains not reactive when triggered by button click)